Isolate contract tests from implementation tests.
