+++ vavr/src/test/java/io/vavr/collection/SumArbitraryValueObjectsTest.java
@@ -1,0 +1,42 @@
+/*  __    __  __  __    __  ___
+ * \  \  /  /    \  \  /  /  __/
+ *  \  \/  /  /\  \  \/  /  /
+ *   \____/__/  \__\____/__/
+ *
+ * Copyright 2014-2017 Vavr, http://vavr.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.vavr.collection;
+
+import org.assertj.core.api.Assertions;
+import org.junit.Test;
+
+import java.util.function.BiFunction;
+
+public abstract class SumArbitraryValueObjectsTest {
+    @Test
+    public void checkSum() throws Exception {
+        Assertions.assertThat(
+                sum(items(), identityElement(), addFunction()))
+                .isEqualTo(expectedSum());
+    }
+
+    protected abstract ExampleSummableValue expectedSum();
+    protected abstract BiFunction<ExampleSummableValue, ExampleSummableValue, ExampleSummableValue> addFunction();
+    protected abstract ExampleSummableValue identityElement();
+    protected abstract List<ExampleSummableValue> items();
+    private ExampleSummableValue sum(List<ExampleSummableValue> items, ExampleSummableValue identityElement, BiFunction<ExampleSummableValue, ExampleSummableValue, ExampleSummableValue> addFunction) {
+        return items.foldLeft(identityElement, addFunction);
+    }
+}
